Update sc_qubits.py
Definition of method `hilbertdim` was missing in class QubitFluxonium. This led to the output `Hilbert space dimension: None` when printing the information for an instance of the class. (Note that `dim` and `hilbertdim` have different meanings. The former defines the number of levels to be kept, e.g., when coupling to an oscillator. The latter states the number of oscillator states used in the diagonalization of the bare fluxonium Hamiltonian.)